fix: repair Supabase pgTAP and web lint CI on migration branch
- Correct delete_all_user_cloud_data test plan count (8 → 9)
- Use integer literals for jsonb_array_length assertions (not bigint)
- Replace spread+sort with toSorted() in notesLocalRepository test

+++ supabase/tests/database/notelikeus_list_user_attachments.test.sql
@@ -22,7 +22,7 @@ select lives_ok(
 
 select results_eq(
   $$ select jsonb_array_length(public.list_user_attachments()) $$,
-  ARRAY[1::bigint],
+  ARRAY[1],
   'list_user_attachments returns active attachments'
 );
